feat: implement analysis persistence layer
- added AnalysisService for saving and retrieving analysis
- implemented AnalysisRepository with user-based queries
- created AnalysisResponseDTO for structured responses
- added mapping from entity to DTO
- implemented sorting by createdAt (latest first)
- integrated JWT-based user identification in service layer

+++ backend/src/main/java/com/project/skillGapAnalyzer/dto/response/SkillAnalysisResponseDTO.java
@@ -16,7 +16,7 @@ public class SkillAnalysisResponseDTO {
 
     private List<String> matchedSkills;
     private List<String> missingSkills;
-    private double progress;
+    private int progress;
     private Map<String, List<String>> recommendedResources;
     private Map<String, List<RepoDTO>> recommendedRepositories;
